Changed: check if column exists before adding or droping

+++ migrations/2025_12_03_000000_add_card_image_url_to_discussions.php
@@ -14,13 +14,19 @@ use Illuminate\Database\Schema\Builder;
 
 return [
     'up' => function (Builder $schema) {
+        if ($schema->hasColumn('discussions', 'walsgit_card_image_url')) {
+            return;
+        }
 
         $schema->table('discussions', function (Blueprint $table) {
-                $table->string('walsgit_card_image_url')->nullable();
+            $table->string('walsgit_card_image_url')->nullable();
         });
     },
 
     'down' => function (Builder $schema) {
+        if (! $schema->hasColumn('discussions', 'walsgit_card_image_url')) {
+            return;
+        }
         $schema->table('discussions', function (Blueprint $table) {
             $table->dropColumn('walsgit_card_image_url');
         });
